fix(compiler): BUG-920/921 — the analyzer read freed memory to decide a safety verdict; add tools/compiler_asan_sweep.sh
Found by a sweep nobody had run: ZERC ITSELF under ASan+UBSan over the whole
corpus. Every gate asks about the program ZER compiles, and ub_sweep.sh /
ubsan_sweep.sh instrument the EMITTED C. Nothing instrumented the compiler.

BUG-920 — heap-use-after-free in zercheck_ir.c, three sites.
ir_add_handle grows ps->handles with realloc, so every pointer into that
array is invalidated by it. That is exactly why IRAliasSnapshot exists, and
its usage note spells out the ordering. Three sites deviated:
  3669  snapshot correct, then `fd->state = fh->state` — a stale read
  3706  same shape; the one ASan reached, via tests/zer/tokenizer.zer
  5417  snapshot taken AFTER the add, so alloc_id, pool_name, escaped and
        the view set were ALL copied out of freed memory
No .zer test could have caught this: reading freed memory does not crash —
the block is still mapped and usually still holds the old bytes — it just
makes the leak / wrong-pool / use-after-free verdict depend on what the
allocator left behind, silently and non-deterministically.

BUG-921 — `1LL << 63` and the INT64_MIN - 1 after it, computing an @saturate
bound in the emitter. GCC wraps, so the emitted C was right by luck; a
compiler free to fold it differently would emit a wrong clamp with no
diagnostic anywhere. Computed unsigned now. The AST-path twin uses
hard-coded literals per width and was never affected.

tools/compiler_asan_sweep.sh is the third sweep and subsumes neither of the
others: ub_sweep is -O0/-O2 differential on emitted C, ubsan_sweep is
sanitizers on emitted C, this is sanitizers on the compiler. Verified
non-vacuous: RED on the pre-fix tree, green after. It builds into a temp dir
so it can never leave instrumented .o files behind — CLAUDE.md's #1
phantom-bug trap.

Sweep after the fix: 2223 inputs, 0 findings.
make check exit 0, 1412 .zer, all nine gates, sink matrix 103/103.

+++ emitter.c
@@ -7989,7 +7989,18 @@ static void emit_rewritten_node(Emitter *e, Node *node, IRFunc *func) {
                      * "integer constant is so large that it is unsigned" warnings (and
                      * `-Werror` failures). Emit `(LLONG_MIN_LITERAL)` for w==64 and
                      * the direct form for smaller widths. */
-                    int64_t max_v = (1LL << (w - 1)) - 1;
+                    /* BUG-921: computed in UNSIGNED. `1LL << 63` shifts into the
+                     * sign bit of a signed 64-bit type — undefined behaviour in
+                     * the COMPILER — and the `- 1` that followed then overflowed
+                     * INT64_MIN. Both were flagged by a UBSan build of zerc on
+                     * tests/zer/audit_saturate_i64_literal_ok.zer. The emitted C
+                     * was right by luck (GCC wraps, landing on INT64_MAX, and the
+                     * w == 64 branch below does not even use max_v for the MIN
+                     * side), which is exactly why it went unnoticed: a compiler
+                     * free to fold this differently would emit a wrong clamp with
+                     * no diagnostic anywhere. Unsigned shift and subtract are
+                     * fully defined at every width from 1 to 64. */
+                    int64_t max_v = (int64_t)((1ULL << (w - 1)) - 1ULL);
                     if (w == 64) {
                         emit(e, "_zer_sat%d < (-9223372036854775807LL - 1) ? (-9223372036854775807LL - 1) : "
                                 "_zer_sat%d > %lldLL ? (%lldLL) : (",
